+++ controller/resource_test.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"testing"
 
+	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ecs"
 	"github.com/aws/aws-sdk-go/service/secretsmanager"
 	"github.com/hashicorp/consul-ecs/controller/mocks"
@@ -24,10 +25,10 @@ func TestTaskLister_List(t *testing.T) {
 	for name, c := range cases {
 		t.Run(name, func(t *testing.T) {
 			task1 := &ecs.Task{
-				TaskArn: pointerToStr("task1"),
+				TaskArn: aws.String("task1"),
 			}
 			task2 := &ecs.Task{
-				TaskArn: pointerToStr("task2"),
+				TaskArn: aws.String("task2"),
 			}
 			tl := TaskLister{ECSClient: &mocks.ECSClient{Tasks: []*ecs.Task{task1, task2}, PaginateResults: c.paginateResults}}
 
@@ -51,59 +52,59 @@ func TestTask_Upsert(t *testing.T) {
 	}{
 		"task without mesh tag": {
 			task: &ecs.Task{
-				TaskArn:           pointerToStr("task-arn"),
-				TaskDefinitionArn: pointerToStr("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
+				TaskArn:           aws.String("task-arn"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
 			},
-			existingSecret:     &secretsmanager.GetSecretValueOutput{Name: pointerToStr("test-service"), SecretString: pointerToStr(`{}`)},
+			existingSecret:     &secretsmanager.GetSecretValueOutput{Name: aws.String("test-service"), SecretString: aws.String(`{}`)},
 			expectTokenToExist: false,
 		},
 		"task with mesh tag": {
 			task: &ecs.Task{
-				TaskArn:           pointerToStr("task"),
-				TaskDefinitionArn: pointerToStr("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
-				Tags:              []*ecs.Tag{{Key: pointerToStr(meshTag), Value: pointerToStr("true")}},
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
+				Tags:              []*ecs.Tag{{Key: aws.String(meshTag), Value: aws.String("true")}},
 			},
-			existingSecret:     &secretsmanager.GetSecretValueOutput{Name: pointerToStr("test-service"), SecretString: pointerToStr(`{}`)},
+			existingSecret:     &secretsmanager.GetSecretValueOutput{Name: aws.String("test-service"), SecretString: aws.String(`{}`)},
 			expectTokenToExist: true,
 		},
 		"task with an invalid task definition ARN (no / separator)": {
 			task: &ecs.Task{
-				TaskArn:           pointerToStr("task"),
-				TaskDefinitionArn: pointerToStr("arn:aws:ecs:us-east-1:1234567890:task-definition-service:1"),
-				Tags:              []*ecs.Tag{{Key: pointerToStr(meshTag), Value: pointerToStr("true")}},
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition-service:1"),
+				Tags:              []*ecs.Tag{{Key: aws.String(meshTag), Value: aws.String("true")}},
 			},
-			existingSecret: &secretsmanager.GetSecretValueOutput{Name: pointerToStr("test-service"), SecretString: pointerToStr(`{}`)},
+			existingSecret: &secretsmanager.GetSecretValueOutput{Name: aws.String("test-service"), SecretString: aws.String(`{}`)},
 			expectedError:  `could not determine service name: cannot determine task family from task definition ARN: "arn:aws:ecs:us-east-1:1234567890:task-definition-service:1"`,
 		},
 		"task with an invalid task definition ARN (no revision)": {
 			task: &ecs.Task{
-				TaskArn:           pointerToStr("task"),
-				TaskDefinitionArn: pointerToStr("arn:aws:ecs:us-east-1:1234567890:task-definition/service"),
-				Tags:              []*ecs.Tag{{Key: pointerToStr(meshTag), Value: pointerToStr("true")}},
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service"),
+				Tags:              []*ecs.Tag{{Key: aws.String(meshTag), Value: aws.String("true")}},
 			},
-			existingSecret: &secretsmanager.GetSecretValueOutput{Name: pointerToStr("test-service"), SecretString: pointerToStr(`{}`)},
+			existingSecret: &secretsmanager.GetSecretValueOutput{Name: aws.String("test-service"), SecretString: aws.String(`{}`)},
 			expectedError:  `could not determine service name: cannot determine task family from task definition ARN: "arn:aws:ecs:us-east-1:1234567890:task-definition/service"`,
 		},
 		"when there is an existing token for the service, we don't create a new one": {
 			task: &ecs.Task{
-				TaskArn:           pointerToStr("task"),
-				TaskDefinitionArn: pointerToStr("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
-				Tags:              []*ecs.Tag{{Key: pointerToStr(meshTag), Value: pointerToStr("true")}},
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
+				Tags:              []*ecs.Tag{{Key: aws.String(meshTag), Value: aws.String("true")}},
 			},
 			// When createExistingToken is true, existingSecret will be updated with the value of the created token.
-			existingSecret:      &secretsmanager.GetSecretValueOutput{Name: pointerToStr("test-service"), SecretString: pointerToStr(`{}`)},
+			existingSecret:      &secretsmanager.GetSecretValueOutput{Name: aws.String("test-service"), SecretString: aws.String(`{}`)},
 			createExistingToken: true,
 			expectTokenToExist:  true,
 		},
 		"when the token in the secret doesn't exist in Consul, the secret is updated with the new value": {
 			task: &ecs.Task{
-				TaskArn:           pointerToStr("task"),
-				TaskDefinitionArn: pointerToStr("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
-				Tags:              []*ecs.Tag{{Key: pointerToStr(meshTag), Value: pointerToStr("true")}},
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
+				Tags:              []*ecs.Tag{{Key: aws.String(meshTag), Value: aws.String("true")}},
 			},
 			existingSecret: &secretsmanager.GetSecretValueOutput{
-				Name:         pointerToStr("test-service"),
-				SecretString: pointerToStr(`{"accessor_id":"aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa","token":"bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb"}`),
+				Name:         aws.String("test-service"),
+				SecretString: aws.String(`{"accessor_id":"aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa","token":"bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb"}`),
 			},
 			expectTokenToExist: true,
 		},
@@ -138,7 +139,7 @@ func TestTask_Upsert(t *testing.T) {
 				require.NoError(t, err)
 				secretValue, err := json.Marshal(tokenSecretJSON{AccessorID: token.AccessorID, Token: token.SecretID})
 				require.NoError(t, err)
-				c.existingSecret.SecretString = pointerToStr(string(secretValue))
+				c.existingSecret.SecretString = aws.String(string(secretValue))
 			}
 
 			taskTokens := Task{
@@ -186,6 +187,161 @@ func TestTask_Upsert(t *testing.T) {
 	}
 }
 
-func TestTask_Delete(t *testing.T) {}
+func TestTask_Delete(t *testing.T) {
+	cases := map[string]struct {
+		createExistingToken                bool
+		updateExistingSecret               bool
+		registerConsulService              bool
+		task                               ecs.Task
+		expectTokenDeletedAndSecretEmptied bool
+	}{
+		"the token for service doesn't exist in consul and the secret is empty": {
+			createExistingToken:  false,
+			updateExistingSecret: false,
+			task: ecs.Task{
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
+				Tags:              []*ecs.Tag{{Key: aws.String(meshTag), Value: aws.String("true")}},
+			},
+			expectTokenDeletedAndSecretEmptied: true,
+		},
+		"the token for service doesn't exist in consul and the secret has some value": {
+			createExistingToken:  false,
+			updateExistingSecret: true,
+			task: ecs.Task{
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
+				Tags:              []*ecs.Tag{{Key: aws.String(meshTag), Value: aws.String("true")}},
+			},
+			expectTokenDeletedAndSecretEmptied: true,
+		},
+		"the token for service exists in consul and the secret has the token value": {
+			createExistingToken:  true,
+			updateExistingSecret: true,
+			task: ecs.Task{
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
+				Tags:              []*ecs.Tag{{Key: aws.String(meshTag), Value: aws.String("true")}},
+			},
+			expectTokenDeletedAndSecretEmptied: true,
+		},
+		"the token for service exists in consul but the secret is empty": {
+			createExistingToken:  true,
+			updateExistingSecret: false,
+			task: ecs.Task{
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
+				Tags:              []*ecs.Tag{{Key: aws.String(meshTag), Value: aws.String("true")}},
+			},
+			expectTokenDeletedAndSecretEmptied: true,
+		},
+		"skips non-mesh services": {
+			createExistingToken:  true,
+			updateExistingSecret: true,
+			task: ecs.Task{
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
+			},
+			expectTokenDeletedAndSecretEmptied: false,
+		},
+		"skips if there is a service still registered with Consul": {
+			createExistingToken:   true,
+			updateExistingSecret:  true,
+			registerConsulService: true,
+			task: ecs.Task{
+				TaskArn:           aws.String("task"),
+				TaskDefinitionArn: aws.String("arn:aws:ecs:us-east-1:1234567890:task-definition/service:1"),
+				Tags:              []*ecs.Tag{{Key: aws.String(meshTag), Value: aws.String("true")}},
+			},
+			expectTokenDeletedAndSecretEmptied: false,
+		},
+	}
 
-func pointerToStr(s string) *string { return &s }
+	for name, c := range cases {
+		t.Run(name, func(t *testing.T) {
+			existingSecret := &secretsmanager.GetSecretValueOutput{Name: aws.String("test-service"), SecretString: aws.String(`{}`)}
+			smClient := &mocks.SMClient{Secret: existingSecret}
+			adminToken := "123e4567-e89b-12d3-a456-426614174000"
+			testServer, err := testutil.NewTestServerConfigT(t, func(c *testutil.TestServerConfig) {
+				c.ACL.Enabled = true
+				c.ACL.Tokens.Master = adminToken
+				c.ACL.DefaultPolicy = "deny"
+			})
+			require.NoError(t, err)
+			defer func() { _ = testServer.Stop() }()
+			testServer.WaitForLeader(t)
+
+			clientConfig := api.DefaultConfig()
+			clientConfig.Address = testServer.HTTPAddr
+			clientConfig.Token = adminToken
+
+			consulClient, err := api.NewClient(clientConfig)
+			require.NoError(t, err)
+
+			// Create existing token in consul and update existing secret.
+			var token *api.ACLToken
+			if c.createExistingToken {
+				token, _, err = consulClient.ACL().TokenCreate(&api.ACLToken{
+					ServiceIdentities: []*api.ACLServiceIdentity{{ServiceName: "service"}},
+				}, nil)
+				require.NoError(t, err)
+			}
+			if c.updateExistingSecret {
+				if token != nil {
+					secretValue, err := json.Marshal(tokenSecretJSON{AccessorID: token.AccessorID, Token: token.SecretID})
+					require.NoError(t, err)
+					existingSecret.SecretString = aws.String(string(secretValue))
+				} else {
+					secretValue, err := json.Marshal(tokenSecretJSON{AccessorID: "some-accessor-id", Token: "some-secret-id"})
+					require.NoError(t, err)
+					existingSecret.SecretString = aws.String(string(secretValue))
+				}
+			}
+
+			anotherToken, _, err := consulClient.ACL().TokenCreate(&api.ACLToken{
+				ServiceIdentities: []*api.ACLServiceIdentity{{ServiceName: "another-service"}},
+			}, nil)
+			require.NoError(t, err)
+
+			if c.registerConsulService {
+				err := consulClient.Agent().ServiceRegister(&api.AgentServiceRegistration{
+					Name: "service",
+				})
+				require.NoError(t, err)
+			}
+
+			taskTokens := Task{
+				SecretsManagerClient: smClient,
+				ConsulClient:         consulClient,
+				Cluster:              "test-cluster",
+				SecretPrefix:         "test",
+				Task:                 c.task,
+				Log:                  hclog.NewNullLogger(),
+			}
+
+			err = taskTokens.Delete()
+			require.NoError(t, err)
+
+			if c.createExistingToken {
+				// Check that the token is deleted from Consul.
+				_, _, err = consulClient.ACL().TokenRead(token.AccessorID, nil)
+				if c.expectTokenDeletedAndSecretEmptied {
+					require.EqualError(t, err, "Unexpected response code: 403 (ACL not found)")
+				} else {
+					require.NoError(t, err)
+				}
+			}
+
+			if c.expectTokenDeletedAndSecretEmptied {
+				// Check that the secret is updated to an empty string.
+				require.Equal(t, `{}`, *smClient.Secret.SecretString)
+			} else {
+				require.Equal(t, *existingSecret.SecretString, *smClient.Secret.SecretString)
+			}
+
+			// Check that the other token is not affected.
+			_, _, err = consulClient.ACL().TokenRead(anotherToken.AccessorID, nil)
+			require.NoError(t, err)
+		})
+	}
+}